requirements.txt
numpy>=1.19
scipy>=1.8
numba
matplotlib
pyyaml